requirements.txt
opencv-python~=4.5.3.56
mss~=6.1.0
numpy~=1.21.1
Pillow~=8.3.1
imutils~=0.5.4
pynput~=1.7.3
matplotlib~=3.4.2